refactor(repo): 抽取 buildClosedSessionCard 去重 + 补 repo_manual_submit 的 operatorOpenId
把「捕获旧会话身份 + 生成 claude --resume 命令 + buildSessionClosedCard」
这段在 /close、两条切库路径、卡片 close 按钮共四处逐字重复的逻辑,抽成
src/core/closed-session-card.ts 的 buildClosedSessionCard(ds, locale),四处
统一调用,去掉 ~90 行重复,也防止四处日后漂移。

顺手补 review 提到的 parity 缺口:repo_manual_submit(手动输入路径切库)
那条 commitRepoSelection 调用的 ctx 漏传 operatorOpenId,导致关闭卡走
群内可见而非 ephemeral 仅自己可见 —— 与下拉/worktree 路径不一致。补上
operatorOpenId 即可对齐。
